Quitar campos checkout según método de envío
Instead of unsetting the fields, we will just alter it's requiredness. That means, if the chosen method is the one we want to check, we will not make it required. Then we will add a hide class. With this, we can hide these fields using css. And woocommerce will not throw an error that it is required. Using jQuery, we can show/hide these fields. So if we unset it on the first run, there's nothing to show because the fields are not there in the first place and with that the page needs to reload.

Source: https://stackoverflow.com/questions/50498391/remove-selected-checkout-fields-with-woocommerce-depending-on-shipping-method